requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.13
django-phonenumber-field==6.3.0
django-rest-framework==0.1.0
django-twilio==0.13.2
djangorestframework==3.13.1
idna==3.3
phonenumbers==8.12.50
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
twilio==6.63.2
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
